Move processes and gdal_functions to geo package; remove old cmake files
