--- cmake/travis_continuous.cmake
@@ -1,26 +1,0 @@
-set(CTEST_SOURCE_DIRECTORY "$ENV{TRAVIS_BUILD_DIR}")
-set(CTEST_BINARY_DIRECTORY "$ENV{TRAVIS_BUILD_DIR}/_build")
-
-include(${CTEST_SOURCE_DIRECTORY}/CTestConfig.cmake)
-set(CTEST_SITE "Travis")
-set(CTEST_BUILD_NAME "Linux-$ENV{TRAVIS_BRANCH}-$ENV{TRAVIS_PYTHON_VERSION}")
-set(CTEST_CMAKE_GENERATOR "Unix Makefiles")
-
-ctest_start("Continuous")
-if($ENV{TRAVIS_PYTHON_VERSION} STREQUAL 2.6)
-  # pycoverage seems broken on 2.6
-  ctest_configure(OPTIONS "-DPYTHON_COVERAGE=OFF")
-else()
-  ctest_configure()
-endif()
-ctest_build()
-ctest_test(PARALLEL_LEVEL 4 RETURN_VALUE res)
-ctest_coverage()
-file(REMOVE "${CTEST_BINARY_DIRECTORY}/coverage.xml")
-ctest_submit()
-
-file(REMOVE "${CTEST_BINARY_DIRECTORY}/test_failed")
-if(NOT res EQUAL 0)
-  file(WRITE "${CTEST_BINARY_DIRECTORY}/test_failed" "error")
-  message(FATAL_ERROR "Test failures occurred.")
-endif()
