+++ gaia/inputs.py
@@ -1,9 +1,11 @@
+import errno
 import json
 import os
-import errno
-import geopandas
-import gdal
 import shutil
+
+import gdal
+import geopandas
+
 try:
     import osr
 except ImportError:
@@ -11,7 +13,7 @@ except ImportError:
 import gaia.formats as formats
 from gaia.core import GaiaException, config
 from gaia.filters import filter_pandas
-from gaia.gdal_functions import gdal_reproject
+from gaia.geo.gdal_functions import gdal_reproject
 
 
 class MissingParameterError(GaiaException):
