--- tests/CMakeLists.txt
@@ -1,6 +1,0 @@
-python_tests_init()
-
-add_python_style_test(python_static_analysis "${PROJECT_SOURCE_DIR}/gaia")
-add_python_style_test(python_test_static_analysis "${PROJECT_SOURCE_DIR}/tests")
-
-add_python_test(geoprocess)
